Xóa các tệp không còn cần thiết cho hệ thống quản lý sinh viên phân tán, bao gồm các tệp cấu hình, kịch bản khởi tạo cơ sở dữ liệu và các lớp không sử dụng. Thêm các tệp mới cho các yêu cầu và phản hồi liên quan đến học phí và sinh viên.

+++ src/Shared/Domain/ValueObject/Scores.cs
@@ -1,16 +1,17 @@
 namespace Shared.Domain.ValueObject;
 
-public record Scores(ushort? AttendanceScore,
-    ushort? MidtermScore,
-    ushort? FinalScore)
+public record Scores(
+    float? AttendanceScore,
+    float? MidtermScore,
+    float? FinalScore)
 {
     public float? AvgScore
     {
         get
         {
-            if(!AttendanceScore.HasValue || !MidtermScore.HasValue || !FinalScore.HasValue)
+            if (!AttendanceScore.HasValue || !MidtermScore.HasValue || !FinalScore.HasValue)
                 return null;
-            
+
             const float attendanceWeight = 0.1f;
             const float midtermWeight = 0.3f;
             const float finalWeight = 0.6f;
